+++ package.json
@@ -8,6 +8,7 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@fortawesome/react-fontawesome": "^0.1.9",
+    "@material-ui/core": "^4.9.9",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -23,6 +24,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
+    "typeface-roboto": "0.0.75",
     "typescript": "^3.8.3"
   },
   "devDependencies": {
@@ -63,8 +65,8 @@
     "build-storybook": "build-storybook -s public",
     "format": "prettier --write \"**/*.+(js|jsx|json|yml|yaml|css|md|vue)\"",
     "prestart": "npm run build:styles",
-    "build:styles": "postcss src/index.pcss -o src/index.css",
-    "watch:styles": "postcss src/index.pcss -o src/index.css -w",
+    "build:styles": "postcss src/components/CardsAgainstHumanity/Board/BoardContainer.pcss -o src/components/CardsAgainstHumanity/Board/BoardContainer.css",
+    "watch:styles": "postcss src/components/CardsAgainstHumanity/Board/BoardContainer.pcss -o src/components/CardsAgainstHumanity/Board/BoardContainer.css -w",
     "changelog": "github_changelog_generator -u nahuelhds -p cards-against-humanity"
   },
   "lint-staged": {
