+++ src/assets/decks/es_AR/black.json
