+++ src/components/CardsAgainstHumanity/Board/BoardContainer.pcss
@@ -3,6 +3,5 @@
 @import "tailwindcss/utilities";
 
 body {
-  @apply bg-gray-500;
 
 }
