+++ src/server.js
@@ -1,6 +1,6 @@
 // src/server.js
 const Server = require("boardgame.io/server").Server;
-const { GameCardsAgainstHumanity } = require("./CardsAgainstHumanity/game");
+const { GameCardsAgainstHumanity } = require("./components/CardsAgainstHumanity/game");
 
 const PORT = process.env.PORT || 8000;
 const server = Server({ games: [GameCardsAgainstHumanity] });
